mail fx addd

+++ src/services/auth/login.js
@@ -1,5 +1,5 @@
 import { API_URL } from "../settings"
-export async function login(user) {
+export async function login(user, history, register) {
   //login
   const res = await fetch(`${API_URL}/auth/local`, {
     method: "post",
@@ -14,6 +14,16 @@ export async function login(user) {
 
   //set token in LocalStorage
   const json = await res.json()
+  //si esta registrando, permitir la obtecion del token
+  if(!register){
+    //si no esta confirmada
+    if(json.user.confirmed!== true) {
+      console.log(json)
+      history.push("/mailconfimation-required")    
+      const error=Error("Mail confirmation required")
+      return error
+    }
+  }
   localStorage.setItem("token", json.jwt)
   return json
 }
